+++ src/main/java/com/smartdine/repository/UserRepository.java
@@ -3,7 +3,8 @@ package com.smartdine.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.smartdine.models.User;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     boolean existsByEmail(String email);
+
     boolean existsByPhone(String phone);
 }
